\name{dat}
\alias{dat}
\docType{data}
\title{
Data for example in \code{\link[gim]{gim}}
}
\description{
\code{dat} is a data frame used in the example of \code{\link[gim]{gim}}. 
}
\usage{data("dat")}
\format{
  A data frame with 4000 observations on the following 6 variables.
  \describe{
    \item{\code{y}}{a continuous outcome}
    \item{\code{d}}{a binary outcome}
    \item{\code{x1}}{a numeric variable}
    \item{\code{x2}}{a numeric variable}
    \item{\code{x3}}{a numeric variable}
    \item{\code{x4}}{a character variable}
  }
}
\details{
This is a dataset from which internal and external data are extracted for the example. 
}

\keyword{datasets}
