\name{fit}
\alias{fit}
\docType{data}
\title{
Retured object of \code{\link[gim]{gim}} in example
}
\description{
\code{fit} is an object fitted as an example. It is saved with the package to save time in illustration. More details about the structure of \code{fit} are given in 'Details' of \code{\link[gim]{gim}}. 
}
\usage{data("fit")}
\format{
  The format is:
List of 5
 $ coefficients: Named num [1:10] 0.7875 0.1669 0.261 -0.1336 -0.0683 ...
  ..- attr(*, "names")= chr [1:10] "(Intercept)" "age_50to59" "age_gt60" "breastfeedyes" ...
 $ vcov        : num [1:10, 1:10] 4.14e-03 -2.97e-03 -3.05e-03 -4.71e-05 -1.94e-04 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "(Intercept)" "age_50to59" "age_gt60" "breastfeedyes" ...
  .. ..$ : chr [1:10] "(Intercept)" "age_50to59" "age_gt60" "breastfeedyes" ...
 $ sigma2      : Named num 0.0476
  ..- attr(*, "names")= chr "sigma"
 $ score       : Named num [1:51] -2.03e-13 1.33e-12 9.74e-11 5.57e-13 3.65e-12 ...
  ..- attr(*, "names")= chr [1:51] "lam1" "lam2" "lam3" "lam4" ...
 $ call        : language gim.default(formula = form, family = "gaussian", data = dat, model = model, nsample = nsample)
 - attr(*, "class")= chr "gim"
}
\source{
This is an artificial dataset generated by function \code{gen.example} in this package. 
}
\keyword{datasets}
