\name{form}
\alias{form}
\docType{data}
\title{
Formula for example in \code{[gim]{gim}}
}
\description{
\code{form} is a general formula used in example. 
}
\usage{data("form")}
\format{
  The format is:
Class 'formula'  language log(density) ~ age_50to59 + age_gt60 + breastfeed + parity + as.factor(p53) * I(subtype ==      "A")
  ..- attr(*, ".Environment")=<environment: R_GlobalEnv> 
}
\source{
This is an artificial dataset generated by function \code{gen.example} in this package. 
}
\keyword{datasets}
