\name{model}
\alias{model}
\docType{data}
\title{
Model of summary information for example in \code{\link[gim]{gim}}
}
\description{
\code{model} is a list specifying the working models in example. 
}
\usage{data("model")}
\format{
  The format is:
List of 2
 $ :List of 2
  ..$ formula: chr "log(density) ~ I(age >= 70) + breastfeed + subtype"
  ..$ info   :'data.frame':	3 obs. of  2 variables:
  .. ..$ var: Factor w/ 3 levels "breastfeedyes",..: 2 1 3
  .. ..$ bet: num [1:3] 0.0287 -0.1352 0.1115
 $ :List of 2
  ..$ formula: chr "log(density) ~ subtype"
  ..$ info   :'data.frame':	2 obs. of  2 variables:
  .. ..$ var: Factor w/ 2 levels "subtypeB","subtypeC": 1 2
  .. ..$ bet: num [1:2] 0.0665 0.0802
}
\source{
This is an artificial dataset generated by function \code{gen.example} in this package. 
}
\keyword{datasets}
