\name{nsample}
\alias{nsample}
\docType{data}
\title{
Sample sizes of external studies for example in \code{\link[gim]{gim}}
}
\description{
\code{nsample} is a squred matrix specifying the number of samples shared by external studies.  
}
\usage{data("nsample")}
\format{
  The format is:
 num [1:4, 1:4] 200 118 0 0 118 118 0 0 0 0 ...
}
\source{
This is an artificial dataset generated by function \code{gen.example} in this package. 
}
\keyword{datasets}
