\name{dat}
\alias{dat}
\docType{data}
\title{
Data for example in \code{\link[gim]{gim}}
}
\description{
\code{dat} is a data frame used in the example of \code{\link[gim]{gim}}. 
}
\usage{data("dat")}
\format{
  A data frame with 350 observations on the following 10 variables.
  \describe{
    \item{\code{subj.id}}{a character vector, subject ID}
    \item{\code{density}}{a numeric vector, outcome}
    \item{\code{bmi}}{a numeric vector, body mass index}
    \item{\code{age}}{a numeric vector, age}
    \item{\code{breastfeed}}{a character vector, breast feed or not}
    \item{\code{parity}}{a numeric vector, number of children}
    \item{\code{p53}}{a numeric vector, biomarker}
    \item{\code{subtype}}{a character vector, biomarker}
    \item{\code{age_50to59}}{an integer vector, age category derived from \code{age}}
    \item{\code{age_gt60}}{an integer vector, age category derived from \code{age}}
  }
}
\details{
This is an example of data of internal study. 
}
\source{
This is an artificial dataset generated by function \code{gen.example} in this package. 
}
\keyword{datasets}
