\name{model}
\alias{model}
\docType{data}
\title{
Model for example in \code{\link[gim]{gim}}
}
\description{
\code{model} is a list specifying the working models in example. 
}
\usage{data("model")}
\format{
  The format is:
List of 4
 $ :List of 2
  ..$ formula: chr "log(density) ~ age + I(parity > 0) + subtype"
  ..$ info   :'data.frame':	3 obs. of  2 variables:
  .. ..$ var: Factor w/ 3 levels "I(parity > 0)TRUE",..: 1 2 3
  .. ..$ bet: num [1:3] -0.2618 0.0164 0.0482
 $ :List of 2
  ..$ formula: chr "log(density) ~ I(age >= 70) + breastfeed + subtype"
  ..$ info   :'data.frame':	3 obs. of  2 variables:
  .. ..$ var: Factor w/ 3 levels "breastfeedyes",..: 2 1 3
  .. ..$ bet: num [1:3] -0.02229 -0.14341 0.00541
 $ :List of 2
  ..$ formula: chr "log(density) ~ as.factor(p53)"
  ..$ info   :'data.frame':	2 obs. of  2 variables:
  .. ..$ var: Factor w/ 2 levels "as.factor(p53)2",..: 1 2
  .. ..$ bet: num [1:2] 0.0975 0.1917
 $ :List of 2
  ..$ formula: chr "log(density) ~ subtype"
  ..$ info   :'data.frame':	2 obs. of  2 variables:
  .. ..$ var: Factor w/ 2 levels "subtypeB","subtypeC": 1 2
  .. ..$ bet: num [1:2] 0.087 0.142
}
\source{
This is an artificial dataset generated by function \code{gen.example} in this package. 
}
\keyword{datasets}
