\name{nsample}
\alias{nsample}
\docType{data}
\title{
Sample sizes of external studies for example in \code{\link[gim]{gim}}
}
\description{
\code{nsample} is a squred matrix specifying the number of samples shared by external studies.
}
\usage{data("nsample")}
\format{
  The format is:
 int [1:2, 1:2] 200 100 100 300
}
\source{
This is an artificial dataset generated by function \code{gen.example} in this package. 
}
\keyword{datasets}
